improve Project Euler - Problem 32

+++ src/problems/p_0033/mod.rs
@@ -18,13 +18,13 @@ fn solve() -> String {
 
     for numerator in 10..100 {
         for denominator in (numerator + 1)..100 {
-            let digits_numerator: [u64; 2] = digits(numerator).collect::<Vec<u64>>().try_into().unwrap();
-            let digits_denominator: [u64; 2] = digits(denominator).collect::<Vec<u64>>().try_into().unwrap();
+            let digits_numerator: [u8; 2] = digits(numerator).collect::<Vec<u8>>().try_into().unwrap();
+            let digits_denominator: [u8; 2] = digits(denominator).collect::<Vec<u8>>().try_into().unwrap();
 
             'outer: for m in 0..2 {
                 for n in 0..2 {
                     if (digits_numerator[m] == digits_denominator[n]) && (digits_numerator[m] != 0) {
-                        let lowest_terms = lowest_common_terms([digits_numerator[(m + 1) % 2], digits_denominator[(n + 1) % 2]]);
+                        let lowest_terms = lowest_common_terms([digits_numerator[(m + 1) % 2] as u64, digits_denominator[(n + 1) % 2] as u64]);
                         if lowest_common_terms([numerator, denominator]) == lowest_terms {
                             result[0] *= lowest_terms[0];
                             result[1] *= lowest_terms[1];
